refactor: change grid sizes and structure (#1800)
* refactor: change grid widths

* feat: implement util to manage grid styles in one place

* refactor: move styles from Layout to Sidebar

* refactor: move Topbar components to its own folder
