+++ packages/website/components/LayoutFullSize.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import { css } from 'emotion';
 import { Flex, Box } from '@contentful/f36-components';
-import { Topbar, TopbarHeight } from './Topbar';
+import { Topbar } from './Topbar';
 import { Footer } from './Footer';
+
+import { TOPBAR_HEIGHT } from '../utils/getGridStyles';
 
 interface Props {
   children: React.ReactNode;
@@ -15,7 +17,7 @@ const styles = {
   }),
   main: css({
     width: '100%',
-    height: `calc(100vh - ${TopbarHeight} - 176px)`,
+    height: `calc(100vh - ${TOPBAR_HEIGHT} - 176px)`,
     marginBottom: '176px',
   }),
   footer: css({
